Add color selection as in the book
The book then proceeds by adding color selection capabilities to the app. This gets very interesting because it actually implements a sliding list of colors controlled with the swipe gesture. The swipe gesture is actually what I might need most in other projects and it feels so nice.

I also adjusted a couple of css to make the thing look perfect on the iPhone 5s, my testing device.

I elected to save the previous example in a separate js file, so that I can have it for future reference.

Next up, emulate scroll inertia just like in native applications. This gets trickily cool.

Signed-off-by: Emanuele <feliziani.emanuele@gmail.com>